Full code refactor with templates.
The C++ version has been fully formatted with templates.

Now there are 2 classes micropp<2> and micropp<3> and the constructors does not
need to pass the dimension as the first parameter anymore.

The rest of the interface works exactly the same way as before.

TODO: Add a newer FORTRAN interface, because the old one was removed to delete
the static member use.